feat: move validate_and_sanitize_search_inputs to decorators
