+++ frappe/decorators.py
@@ -1,4 +1,7 @@
+import functools
+
 import frappe
+from frappe.utils.data import cint
 
 
 def whitelist(allow_guest=False, xss_safe=False, methods=None):
@@ -46,3 +49,21 @@ def whitelist(allow_guest=False, xss_safe=False, methods=None):
 		return method or fn
 
 	return innerfn
+
+
+def validate_and_sanitize_search_inputs(fn):
+	@functools.wraps(fn)
+	def wrapper(*args, **kwargs):
+		from frappe.desk.search import sanitize_searchfield
+
+		kwargs.update(dict(zip(fn.__code__.co_varnames, args, strict=False)))
+		sanitize_searchfield(kwargs["searchfield"])
+		kwargs["start"] = cint(kwargs["start"])
+		kwargs["page_len"] = cint(kwargs["page_len"])
+
+		if kwargs["doctype"] and not frappe.db.exists("DocType", kwargs["doctype"]):
+			return []
+
+		return fn(**kwargs)
+
+	return wrapper
